+++ NEWS.md
@@ -5,10 +5,12 @@ This is a major release that updates the internal data model to Camtrap DP 1.0 a
 - `read_camtrap_dp()` (with underscore) is now deprecated. Use `read_camtrapdp()`, which reexports the read functionality from [camtrapdp](https://inbo.github.io/camtrapdp/reference/read_camtrapdp.html) (#298) (#298).
 - Function `get_record_table()` is now deprecated. Use `camtrapR_recordTable()` (#240).
 - Function `get_cam_op()` is now deprecated. Use `camtrapR_cameraOperation()` (#239).
+- Function `get_species()` is now deprecated. Use `taxa()` (#343).
 - Function `get_n_species()` is now deprecated. Use `n_species()` (#243).
 - Function `map_dep()` is now deprecated. Use `map_deployments()` (#231).
-- Filter predicates (`apply_filter_predicate()`, `pred()` `pred_not()` `pred_gt()`, `pred_and()`, `pred_or()`, etc.) are now defunct. Use `filter_deployments()` (#316).
 - Arguments `sex` and `life_stage` are now deprecated in (deprecated) `map_dep()`. Use `filter_observations()` beforehand.
+- The functions `get_scientific_name()` and `check_species()` are now defunct. Use `taxa()` for an overview of all species (#235).
+- Filter predicates (`apply_filter_predicate()`, `pred()` `pred_not()` `pred_gt()`, `pred_and()`, `pred_or()`, etc.) are now defunct. Use `filter_deployments()` (#316).
 - The deprecated argument `datapkg` in `read_camtrap_dp()` is removed.
 - `write_dwc()` has moved to [camtrapdp](https://inbo.github.io/camtrapdp/reference/write_dwc.html). `write_eml()` is not needed for GBIF processing of Camtrap DPs and has been removed ([camtrapdp#61](https://github.com/inbo/camtrapdp/issues/61)).
 - `round_coordinates()` has moved to [camtrapdp](https://inbo.github.io/camtrapdp/reference/round_coordinates.html) (#327).
@@ -20,6 +22,7 @@ This is a major release that updates the internal data model to Camtrap DP 1.0 a
 - The first argument of many functions accepts a Camtrap DP object: not `package` anymore, but `x` (#324).
 - Functions `filter_deployments()`, `filter_observations()` and `filter_media()` allow to filter at data package level (#315). They reexport the functionalities from camtrapdp. See [`camtrapdp::filter_deployments()`](https://inbo.github.io/camtrapdp/reference/filter_deployments.html), [`camtrapdp::filter_observations()`](https://inbo.github.io/camtrapdp/reference/filter_observations.html) and [`camtrapdp::filter_media()`](https://inbo.github.io/camtrapdp/reference/filter_media.html).
 - Function `filter_out_timelapse()` allows to filter out timelapse observations (#306). See [`camtrapdp::filter_out_timelapse()`](https://inbo.github.io/camtrapdp/reference/filter_out_timelapse.html).
+- File `CITATION.cff` added to the package (#345).
 
 # camtraptor 0.26.0